Create Stage_1
1st Stage of the game to be delivered.
30-Sep-2018

Background and Image tool have been implemented in the source.
As well as the first steps in game design creation, where loading
items from a text file representing the level stage. However, no rendering yet.

Next Stage goals, the ability to render the level, and include a hero object to 
test finishing a level and see object being rendered and interacted with.